requirements.txt
flask
Flask-WTF
flask-restful
flask-apispec
WTForms
flask_sqlalchemy
pymysql
requests
pyjwt